Add cleanup_temp_zip_files setting to SSAO.cfg to speed up ZIP load times.
Setting:

	cleanup_temp_zip_files = 0

in SSAO.cfg will prevent temporary unzipped files from being deleted. This will
make load times slightly better (~6s) after the first time at the cost of
increased disk space usage. To revert this, just set this to 1.

(cherry picked from commit 343aa5d87621fe2ca551941c6d718275ca804a79)

+++ impl11/ddraw/DeviceResources.cpp
@@ -1629,7 +1629,7 @@ HRESULT DeviceResources::OnSizeChanged(HWND hWnd, DWORD dwWidth, DWORD dwHeight)
 	ResetObjectIndexMap();
 	ReloadInterdictionMap();
 	//ResetRawMouseInput();
-	if (IsZIPReaderLoaded())
+	if (IsZIPReaderLoaded() && g_bCleanupZIPDirs)
 		DeleteAllTempZIPDirectories();
 	this->ResetExtraTextures();
 	this->_depthStencilViewL.Release();
